remove gulp script 'karma', use updated 'watch' script instead
